RT-DETR/requirements.txt
anyio==4.2.0
appnope==0.1.3
astor==0.8.1
asttokens==2.4.1
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.3.2
coloredlogs==15.0.1
comm==0.2.1
contourpy==1.2.0
cycler==0.10.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
executing==2.0.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.47.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
humanfriendly==10.0
idna==2.10
imageio==2.33.1
imgaug==0.4.0
ipykernel==6.29.0
ipython==8.21.0
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.6
motmetrics==1.4.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.3
onnx==1.13.0
onnxruntime==1.15.1
opencv-python-headless==4.8.0.74
openvino==2023.1.0
openvino-telemetry==2023.2.1
opt-einsum==3.3.0
packaging==23.2
paddle-bfloat==0.1.7
paddledet==0.0.0
paddlepaddle==2.5.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyclipper==1.3.0.post5
pycocotools==2.0.7
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-magic==0.4.27
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
requests-toolbelt==1.0.0
roboflow==1.1.18
scipy==1.12.0
six==1.16.0
sklearn==0.0
sniffio==1.3.0
stack-data==0.6.3
supervision==0.18.0
sympy==1.12
torch==2.0.1
torchvision==0.15.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.9.0
urllib3==2.2.0
wcwidth==0.2.13
